Refactor each class to be in its own file, change entry point to __init__.py
